Bruker rxjs for å styre abb på posisjonsdata
